+++ src/search/ir.h
@@ -33,83 +33,17 @@
 #include <vector>
 
 #include "fmt/core.h"
+#include "ir_iterator.h"
 #include "string_util.h"
 #include "type_util.h"
 
 // kqir stands for Kvorcks Query Intermediate Representation
 namespace kqir {
 
-struct Node;
-
-struct NodeIterator {
-  std::variant<Node *, std::array<Node *, 2>,
-               std::pair<Node *, std::vector<std::function<Node *(Node *)>>::const_iterator>,
-               std::vector<std::unique_ptr<Node>>::iterator>
-      val;
-
-  NodeIterator() : val(nullptr) {}
-  explicit NodeIterator(Node *node) : val(node) {}
-  NodeIterator(Node *n1, Node *n2) : val(std::array<Node *, 2>{n1, n2}) {}
-  explicit NodeIterator(Node *parent, std::vector<std::function<Node *(Node *)>>::const_iterator iter)
-      : val(std::make_pair(parent, iter)) {}
-  template <typename Iterator,
-            std::enable_if_t<std::is_base_of_v<Node, typename Iterator::value_type::element_type>, int> = 0>
-  explicit NodeIterator(Iterator iter) : val(*CastToNodeIter(&iter)) {}
-  
-  template <typename Iterator>
-  static auto CastToNodeIter(Iterator *iter) {
-    auto res __attribute__((__may_alias__)) = reinterpret_cast<std::vector<std::unique_ptr<Node>>::iterator *>(iter);
-    return res;
-  }
-
-  template <auto F>
-  static Node *MemFn(Node *parent) {
-    return (reinterpret_cast<typename GetClassFromMember<decltype(F)>::type *>(parent)->*F).get();
-  }
-
-  friend bool operator==(NodeIterator l, NodeIterator r) {
-    return l.val == r.val;
-  }
-
-  friend bool operator!=(NodeIterator l, NodeIterator r) {
-    return l.val != r.val;
-  }
-
-  Node *operator*() {
-    if (val.index() == 0) {
-      return std::get<0>(val);
-    } else if (val.index() == 1) {
-      return std::get<1>(val)[0];
-    } else if (val.index() == 2) {
-      auto &[parent, iter] = std::get<2>(val);
-      return (*iter)(parent);
-    } else {
-      return std::get<3>(val)->get();
-    }
-  }
-
-  NodeIterator &operator++() {
-    if (val.index() == 0) {
-      val = nullptr;
-    } else if (val.index() == 1) {
-      val = std::get<1>(val)[1];
-    } else if (val.index() == 2) {
-      ++std::get<2>(val).second;
-    } else {
-      ++std::get<3>(val);
-    }
-
-    return *this;
-  }
-};
-
 struct Node {
   virtual std::string Dump() const = 0;
   virtual std::string_view Name() const = 0;
-  virtual std::string DisplayName() {
-    auto name = Name();
-    return {name.begin(), name.end()};
-  }
+  virtual std::string Content() const { return {}; }
 
   virtual NodeIterator ChildBegin() { return {}; };
   virtual NodeIterator ChildEnd() { return {}; };
@@ -142,6 +76,7 @@ struct FieldRef : Node {
 
   std::string_view Name() const override { return "FieldRef"; }
   std::string Dump() const override { return name; }
+  std::string Content() const override { return Dump(); }
 };
 
 struct StringLiteral : Node {
@@ -151,6 +86,7 @@ struct StringLiteral : Node {
 
   std::string_view Name() const override { return "StringLiteral"; }
   std::string Dump() const override { return fmt::format("\"{}\"", util::EscapeString(val)); }
+  std::string Content() const override { return Dump(); }
 };
 
 struct QueryExpr : Node {};
@@ -178,6 +114,7 @@ struct NumericLiteral : Node {
 
   std::string_view Name() const override { return "NumericLiteral"; }
   std::string Dump() const override { return fmt::format("{}", val); }
+  std::string Content() const override { return Dump(); }
 };
 
 // NOLINTNEXTLINE
@@ -247,6 +184,7 @@ struct NumericCompareExpr : BoolAtomExpr {
 
   std::string_view Name() const override { return "NumericCompareExpr"; }
   std::string Dump() const override { return fmt::format("{} {} {}", field->Dump(), ToOperator(op), num->Dump()); };
+  std::string Content() const override { return ToOperator(op); }
 
   NodeIterator ChildBegin() override { return {field.get(), num.get()}; };
   NodeIterator ChildEnd() override { return {}; };
@@ -259,6 +197,7 @@ struct BoolLiteral : BoolAtomExpr {
 
   std::string_view Name() const override { return "BoolLiteral"; }
   std::string Dump() const override { return val ? "true" : "false"; }
+  std::string Content() const override { return Dump(); }
 };
 
 struct QueryExpr;
@@ -311,6 +250,7 @@ struct Limit : Node {
 
   std::string_view Name() const override { return "Limit"; }
   std::string Dump() const override { return fmt::format("limit {}, {}", offset, count); }
+  std::string Content() const override { return fmt::format("{}, {}", offset, count); }
 };
 
 struct SortBy : Node {
@@ -323,6 +263,7 @@ struct SortBy : Node {
 
   std::string_view Name() const override { return "SortBy"; }
   std::string Dump() const override { return fmt::format("sortby {}, {}", field->Dump(), OrderToString(order)); }
+  std::string Content() const override { return OrderToString(order); }
 
   NodeIterator ChildBegin() override { return NodeIterator(field.get()); };
   NodeIterator ChildEnd() override { return {}; };
@@ -350,6 +291,7 @@ struct IndexRef : Node {
 
   std::string_view Name() const override { return "IndexRef"; }
   std::string Dump() const override { return name; }
+  std::string Content() const override { return Dump(); }
 };
 
 struct SearchStmt : Node {
